Folders, priority, remove checked task user setting, show time instead of date user setting
