+++ components/settings/General.js
@@ -18,6 +18,8 @@ export default function General() {
 
     const { theme, setTheme } = useTheme()
     const [checkboxThemeSync, setCheckboxThemeSync] = useState(false)
+    const [removeCheckedTasks, setRemoveCheckedTasks] = useState()
+    const [showTimeForTasks, setShowTimeForTasks] = useState()
     const [timeZone, setTimeZone] = useState('')
     const [firstDayOfWeek, setFirstDayOfWeek] = useState('')
 
@@ -31,7 +33,7 @@ export default function General() {
     async function GetData() {
         const { data, error } = await supabase
             .from('profiles')
-            .select(`syncTheme , timeZone, firstDayOfWeek, mode`)
+            .select(`syncTheme , timeZone, firstDayOfWeek, mode, removeCheckedTasks, showTimeForTasks`)
             .eq('id', user.id)
             .single()
 
@@ -41,6 +43,8 @@ export default function General() {
         if (data) {
             setCheckboxThemeSync(data.syncTheme)
             setTimeZone(data.timeZone)
+            setRemoveCheckedTasks(data.removeCheckedTasks)
+            setShowTimeForTasks(data.showTimeForTasks)
             console.log(data.timeZone)
             // setFirstDayOfWeek(data.firstDayOfWeek)
             // setTheme(data.mode)
@@ -101,6 +105,36 @@ export default function General() {
         }
 
 
+    }
+
+    //update removeCheckedTasks in database
+    async function UpdateRemoveCheckedTasks(e) {
+        const { data, error } = await supabase
+            .from('profiles')
+            .update({ removeCheckedTasks: e })
+            .eq('id', user.id)
+
+        if (error) {
+            console.log(error)
+        }
+        else {
+            console.log(data)
+        }
+    }
+
+    // update showTimeForTasks
+    async function UpdateShowTimeForTasks(e) {
+        const { data, error } = await supabase
+            .from('profiles')
+            .update({ showTimeForTasks: e })
+            .eq('id', user.id)
+
+        if (error) {
+            console.log(error)
+        }
+        else {
+            console.log(data)
+        }
     }
 
     function lg() {
@@ -196,6 +230,32 @@ export default function General() {
                     <label htmlFor="toggle" className={styles.toggleLabel}></label>
                 </div>
 
+            </div>
+
+            <div className={styles.optionsHorizontal}>
+                <div className={styles.details}>
+                    <div className={styles.minititle}>Automatically removing checked tasks</div>
+                    <div className={styles.discription}>Automatically remove tasks when you click on their checkbox</div>
+                </div>
+                <div className={styles.toggleContainer}>
+                    <input type="checkbox" name="removetoggle" id="removetoggle" className={styles.toggleInput}
+                        onChange={() => { setRemoveCheckedTasks(!removeCheckedTasks), UpdateRemoveCheckedTasks(!removeCheckedTasks) }} checked={removeCheckedTasks}
+                    />
+                    <label htmlFor="removetoggle" className={styles.toggleLabel}></label>
+                </div>
+            </div>
+
+            <div className={styles.optionsHorizontal}>
+                <div className={styles.details}>
+                    <div className={styles.minititle}>Show time for tasks</div>
+                    <div className={styles.discription}>Show the time for tasks instead of the date</div>
+                </div>
+                <div className={styles.toggleContainer}>
+                    <input type="checkbox" name="showTimeToggle" id="showTimeToggle" className={styles.toggleInput}
+                        onChange={() => { setShowTimeForTasks(!showTimeForTasks), UpdateShowTimeForTasks(!showTimeForTasks) }} checked={showTimeForTasks}
+                    />
+                    <label htmlFor="showTimeToggle" className={styles.toggleLabel}></label>
+                </div>
             </div>
             <Toaster position="bottom-right" reverseOrder={false} />
