+++ pages/app/index.js
@@ -7,6 +7,7 @@ import AppLayout from "/layouts/appLayout"
 import LoadingLine from '/components/loadingLine'
 import { SettingsProvider } from "/layouts/stateStore"
 
+import ListOfTasks from "/components/listOfTasks"
 import Calendar from '/components/calendar'
 export default function IndexApp() {
 
@@ -28,31 +29,6 @@ export default function IndexApp() {
             // router.push("/sign-in")
         }
     })
-
-    useEffect(() => {
-        getAllTask()
-    }, [todoPanelToggle])
-
-
-    //get all task from supabase
-    async function getAllTask() {
-        const { data, error } = await supabase
-            .from('todos')
-            .select('*')
-            .order('title', { ascending: true })
-
-        if (data) {
-            setData(data)
-        }
-
-        if (error) {
-            console.log(error)
-        }
-
-        else {
-            console.log(data)
-        }
-    }
 
     if (session) {
         return (
@@ -76,18 +52,9 @@ export default function IndexApp() {
                             <div className={styles.todoList}>
 
                                 <div className={styles.todoListHeader}>
-                                    <h3>Todo List</h3>
+                                    <h3>Tasks List</h3>
 
-                                    {/*  list of tasks  */}
-                                    {data.map((item) => {
-                                        return (
-                                            <div className={styles.todoItem} key={item.id}>
-                                                <h4>{item.title}</h4>
-                                                <p>{item.description}</p>
-                                            </div>
-                                        )
-                                    }
-                                    )}
+                                    <ListOfTasks></ListOfTasks>
 
                                 </div>
                             </div>
