+++ pages/app/tasks/index.js
@@ -1,0 +1,281 @@
+import React, { useEffect, useState } from 'react'
+import { useUser, useSupabaseClient, useSession } from '@supabase/auth-helpers-react'
+import styles from "/styles/task.module.scss"
+import AppLayout from "/layouts/appLayout"
+import ListOfTasks from "/components/listOfTasks"
+import { SettingsProvider } from "/layouts/stateStore"
+
+export default function Tasks() {
+    const session = useSession()
+    const supabase = useSupabaseClient()
+    const user = useUser()
+
+    // Get all folders from supabase
+    async function GetFoldersSupabase() {
+        const { data, error } = await supabase
+            .from('todosFolders')
+            .select('*')
+        if (data) {
+            setFolders(data)
+        }
+        if (error) {
+            console.log('error', error)
+            return
+        }
+    }
+
+    // Create a new folder in supabase
+    async function CreateFolderSupabase() {
+        const { data, error } = await supabase
+                .from('todosFolders')
+                .insert({ user_id: user.id, title: folderTitle})
+    }
+
+    // Edits a folder in supabase
+    async function EditFolderSupabase() {
+        const { data, error } = await supabase
+                .from('todosFolders')
+                .update({title: folderTitle})
+                .eq('id', 3)
+    }
+
+    // Create a new task in supabase
+    async function CreateTaskSupabase() {
+        console.log("FolderID: " + folderID)
+        let nullDate = new Date(null)
+        if (taskDate.getFullYear === nullDate.getFullYear && taskTime.getFullYear === nullDate.getFullYear) {
+            const { data, error } = await supabase
+                .from('todos')
+                .insert({ user_id: user.id, title: taskTitle, date: null, time: null, description: taskDescription, priority: taskPriority, folder_id: folderID})
+        }
+        else if (taskDate.getFullYear === nullDate.getFullYear) {
+            const { data, error } = await supabase
+                .from('todos')
+                .insert({ user_id: user.id, title: taskTitle, date: null, time: taskTime, description: taskDescription, priority: taskPriority, folder_id: folderID })
+        }
+        else if (taskTime.getFullYear === nullDate.getFullYear) {
+            const { data, error } = await supabase
+                .from('todos')
+                .insert({ user_id: user.id, title: taskTitle, date: taskDate, time: null, description: taskDescription, priority: taskPriority, folder_id: folderID })
+        }
+        else {
+            const { data, error } = await supabase
+                .from('todos')
+                .insert({ user_id: user.id, title: taskTitle, date: taskDate, time: taskTime, description: taskDescription, priority: taskPriority, folder_id: folderID })
+        }
+        setReload(true)
+        setCreatingTask(false)
+    }
+
+    // Edits a task in supabase
+    async function EditTaskSupabase() {
+        if (taskDate === undefined && taskTime === undefined) {
+            const { data, error } = await supabase
+                .from('todos')
+                .update({ title: taskTitle, date: null, time: null, description: taskDescription, priority: taskPriority, folder_id: folderID })
+                .eq('id', taskID)
+        }
+        else if (taskDate === undefined) {
+            const { data, error } = await supabase
+                .from('todos')
+                .update({ title: taskTitle, date: null, time: taskTime, description: taskDescription, priority: taskPriority, folder_id: folderID })
+                .eq('id', taskID)
+        }
+        else if (taskTime === undefined) {
+            const { data, error } = await supabase
+                .from('todos')
+                .update({ title: taskTitle, date: taskDate, time: null, description: taskDescription, priority: taskPriority, folder_id: folderID })
+                .eq('id', taskID)
+        }
+        else {
+            const { data, error } = await supabase
+                .from('todos')
+                .update({ title: taskTitle, date: taskDate, time: taskTime, description: taskDescription, priority: taskPriority, folder_id: folderID })
+                .eq('id', taskID)
+        }
+        setReload(true)
+        setEditingTask(false)
+    }
+
+    // Deletes a task in supabase
+    async function DiscardTaskSupabase() {
+        const { data, error } = await supabase
+            .from('todos')
+            .delete()
+            .eq('id', taskID)
+
+        setReload(true)
+        setEditingTask(false)
+    }
+
+    // Function that sets all the states according to the given task and sets editingTask to true
+    function EditTask(task) {
+        setTaskID(task.id)
+        setTaskTitle(task.title)
+        if (task.date === null) {
+            setTaskDate(new Date(null))
+        }
+        else {
+            setTaskDate(task.date)
+        }
+        if (task.time === null) {
+            setTaskTime(new Date(null))
+        }
+        else {
+            setTaskTime(task.time)
+        }
+        setTaskDescription(task.description)
+        setTaskPriority(0)
+        setEditingTask(true)
+    }
+
+    // Function that resets all the states to their zero state and sets creatingTask to true
+    function CreateTask() {
+        setTaskTitle("")
+        setTaskDate(new Date(null))
+        setTaskTime(new Date(null))
+        setTaskDescription("")
+        setTaskPriority(0)
+        setCreatingTask(true)
+        setReload(false)
+    }
+
+    function CancelTask() {
+        setCreatingTask(false)
+        setEditingTask(false)
+        setReload(false)
+    }
+
+    const [creatingTask, setCreatingTask] = useState(false)
+    const [editingTask, setEditingTask] = useState(false)
+    const [reload, setReload] = useState(false)
+
+    const [taskID, setTaskID] = useState(null)
+    const [taskTitle, setTaskTitle] = useState("")
+    const [taskDate, setTaskDate] = useState(new Date(null))
+    const [taskTime, setTaskTime] = useState(new Date(null))
+    const [taskDescription, setTaskDescription] = useState("")
+    const [taskPriority, setTaskPriority] = useState(0)
+    const [folderID, setFolderID] = useState(null)
+
+    const [folderTitle, setFolderTitle] = useState("Folder Title")
+
+    const [openPanel, setOpenPanel] = useState(false)
+    const [AIPanel, setAIPanel] = useState(false)
+
+    const [folders, setFolders] = useState([])
+
+    useEffect(() => {
+        GetFoldersSupabase()
+    }, [])
+
+    if (session) {
+        return (
+            <div>
+                <div className={creatingTask || editingTask ? styles.blur : null}>
+                    <SettingsProvider>
+                        <AppLayout>
+                            <div className={styles.flexContainer}>
+                                <div>
+                                    <div>
+                                        <button className={styles.mainButton} onClick={CreateTask}>
+                                            <span className={styles.create}>Create </span>
+                                            <span>+</span>
+                                        </button>
+                                        <div className={openPanel ? `${styles.panel} ${styles.openFolderPanel}` : `${styles.panel} ${styles.closedFolderPanel}`}>
+                                            <button className={styles.button} onClick={() => setOpenPanel(!openPanel)}>
+                                            {openPanel ? "Close" : "Open"}
+                                            </button>
+                                            <div>
+                                            {folders.map((folder, i) => (
+                                                <h1 className={styles.h1}>{folder.title}</h1>
+                                            ))}
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                                <div>
+                                    <h1 className={styles.h1}>Tasks</h1>
+                                    <ListOfTasks reload={reload}></ListOfTasks>
+                                </div>
+                                <div className={AIPanel ? `${styles.panel} ${styles.openAIPanel}` : `${styles.panel} ${styles.closedAIPanel}`}>
+                                        <button className={`${styles.AIButton} ${styles.button}`} onClick={() => setAIPanel(!AIPanel)}>
+                                        {AIPanel ? "Close" : "AI Assistance"}
+                                        </button>
+                                </div>
+                            </div>
+                        </AppLayout>
+                    </SettingsProvider>
+                </div>
+
+                <div className={creatingTask || editingTask ? styles.taskForm : `${styles.hiddenTaskForm} ${styles.taskForm}`}>
+                    <button className={styles.cancelButton} onClick={CancelTask}>x</button>
+                    <div className={styles.titleHeader}>
+                        <div className={`${styles.titleInputContainer} ${styles.inputContainer}`}>
+                            <label className={styles.label}>Task title:</label>
+                            <input className={styles.input}
+                                id="title"
+                                type="text"
+                                name="taskTitle"
+                                placeholder="Task Title"
+                                value={taskTitle}
+                                required
+                                onChange={(event) => setTaskTitle(event.target.value)}
+                            ></input>
+                        </div>
+                    </div>
+                    <div className={styles.NonTitleInputContainer}>
+                        <select className={styles.select} onChange={(event) => setTaskPriority(event.target.value)}>
+                            <option value="4">None</option>
+                            <option value="1">1</option>
+                            <option value="2">2</option>
+                            <option value="3">3</option>
+                        </select>
+                        <select className={styles.select} onChange={(event) => setFolderID(event.target.value)}>
+                            <option value="Null">None</option>
+                            <option value="1">Folder 1</option>
+                        </select>
+                        <div className={styles.inputContainer}>
+                            <label className={styles.label}>Date:</label>
+                            <input className={styles.input}
+                                id="date"
+                                type="date"
+                                name="taskDate"
+                                value={taskDate}
+                                onChange={(event) => setTaskDate(event.target.value)}
+                            ></input>
+                        </div>
+                        <div className={styles.inputContainer}>
+                            <label className={styles.label}>Time:</label>
+                            <input className={styles.input}
+                                id="time"
+                                type="time"
+                                name="taskTime"
+                                value={taskTime}
+                                onChange={(event) => setTaskTime(event.target.value)}
+                            ></input>
+                        </div>
+                        <div className={styles.inputContainer}>
+                            <label className={styles.label}>Description:</label>
+                            <textarea className={styles.input}
+                                style={{ height: '120px' }}
+                                rows="10"
+                                cols="10"
+                                id="description"
+                                type="text"
+                                name="taskDescription"
+                                placeholder="Description"
+                                value={taskDescription}
+                                onChange={(event) => setTaskDescription(event.target.value)}
+                            ></textarea>
+                        </div>
+                    </div>
+                    <div styles={styles.buttonContainer}>
+                        <button className={styles.saveButton} onClick={taskTitle.length >= 4 ? (editingTask ? EditTaskSupabase : CreateTaskSupabase) : null}>Save</button>
+                        <button className={`${styles.discardButton} ${styles.saveButton}`} onClick={editingTask ? DiscardTaskSupabase : CancelTask}>Discard</button>
+                    </div>
+                </div>
+            </div>
+        )
+    }
+}
