+++ styles/task.module.scss
@@ -14,6 +14,12 @@
   display: none;
 }
 
+.flexContainer {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
+
 .select {
   height: 28px;
   font-size: medium;
@@ -29,9 +35,9 @@
 }
 
 .mainButton {
-  float: right;
+  float: left;
   margin-top: 10px;
-  margin-right: 20px;
+  margin-left: 20px;
   width: auto;
   min-width: 45px;
   height: 45px;
@@ -50,6 +56,27 @@
 .mainButton:hover {
   background-color: #3b5760;
   transition: background-color 0.3s ease-in-out;
+}
+
+.button {
+  background: var(--special-3);
+  border: 1px solid var(--snowflake-100);
+  border-radius: 6px;
+  font-weight: 700;
+  font-size: 14px;
+  margin-top: 15px;
+  margin-left: 5px;
+  height: 30px;
+  z-index: 0;
+  align-items: center;
+}
+
+.AIButton {
+  margin-right: 20px;
+  position: absolute;
+  right: 0;
+  padding-right: 16px;
+  z-index: 2;
 }
 
 .saveButton {
@@ -118,7 +145,7 @@
 }
 
 .taskForm {
-  position: fixed;
+  position: absolute;
   display: -ms-inline-grid;
   top: 45%;
   left: 50%;
@@ -229,11 +256,32 @@
   transition: 0.8s;
   transition-timing-function: line-height;
 }
-.leftTaskpart {
+
+.taskExtended {
+  border-bottom-left-radius: 0px;
+  border-bottom-right-radius: 0px;
+}
+
+.taskExtension {
+  margin-top: 0px;
+  border-top-left-radius: 0px;
+  border-top-right-radius: 0px;
+  transition: all 0.5s cubic-bezier(0, 0, 0, 0); 
+}
+
+.taskCollapsed {
+  opacity: 0;
+  height: 0px;
+  transition: opacity height 0.5s;
+  transition-timing-function: ease-in;
+}
+
+.taskPart {
   display: flex;
   flex-direction: row;
   align-items: center;
 }
+
 .taskChecked {
   opacity: 30%;
   transition: 0.9s;
@@ -257,6 +305,18 @@
 
   transition: 0.7s;
   transition-timing-function: line-height;
+}
+
+.taskCheckBox1 {
+  border-color: var(--anarchy-600) 
+}
+
+.taskCheckBox2 {
+  border-color: var(--heat-wave-600) 
+}
+
+.taskCheckBox3 {
+  border-color: var(--blue-sparkle-600)
 }
 
 .taskCheckBoxChecked {
@@ -296,14 +356,31 @@
 }
 
 .lineChecked {
-  width: 90%;
+  width: 98%;
 }
 
 .taskDate {
-  float: right;
+  margin-top: 5px;
+  margin-right: 5px;
   color: var(--black);
   font-size: x-small;
   text-align: right;
+}
+
+.taskExtendedDate {
+  margin-right: 17px;
+}
+
+.taskDescription {
+  width: 85%;
+}
+
+.operatorSign {
+  color: var(--black);
+  font-size: x-large;
+  font-weight: bolder;
+  text-align: right;
+  cursor: pointer;
 }
 
 .plus {
@@ -315,4 +392,31 @@
   margin-right: 10px;
   margin-top: 14px;
   cursor: pointer;
+}
+
+.panel {
+    background: var(--special-3);
+    flex-basis: 420px;
+    flex-grow: 0;
+    flex-shrink: 0;
+}
+
+.openFolderPanel {
+  margin-right: 0px;
+  transition: all 0.2s ease-in-out;
+}
+
+.closedFolderPanel {
+  margin-right: 100px;
+  transition: all 0.2s ease-in-out;
+}
+
+.openAIPanel {
+  margin-right: 0px;
+  transition: all 0.2s ease-in-out;
+}
+
+.closedAIPanel {
+  margin-right: -420px;
+  transition: all 0.2s ease-in-out;
 }
